+++ adrs/0007-domain-name.md
@@ -10,4 +10,4 @@ status: accepted
 The [guidance on gov.uk](https://www.gov.uk/guidance/get-an-api-domain-on-govuk) specifies that an API domain should be between 3 and 63 characters long and contain only alphanumeric characters (0-9 and a-z) and the ‘-’ (dash) symbol.
 
 ## Decision
-Add a rule to check that the domain follows our guidance.
+Add rules to check that the domain follows our guidance on length and characters.
