+++ spectral-ruleset-govuk-public/test/testdata/domain/characters/invalid-special-character.yaml
@@ -4,4 +4,4 @@ info:
   version: '1.0.0'
 paths: {}
 servers:
-  - url: 'vehicle-registration-number-*.api.gov.uk'
+  - url: 'https://vehicle-registration-number-*.api.gov.uk'
