+++ spectral-ruleset-govuk-public/test/testdata/domain/length/invalid-too-long.yaml
@@ -1,0 +1,7 @@
+openapi: 3.1.0
+info:
+  title: ''
+  version: '1.0.0'
+paths: {}
+servers:
+  - url: 'https://vehicle-registration-number-vehicle-registration-number-vehicles.api.gov.uk'
