+++ spectral-ruleset-govuk-public/test/testdata/domain/length/valid-multiple-labels.yaml
@@ -1,0 +1,7 @@
+openapi: 3.1.0
+info:
+  title: ''
+  version: '1.0.0'
+paths: {}
+servers:
+  - url: 'https://vehicle-registration-number.123456789012345678901234567890123456789012345678901234567890123.api.gov.uk'
